Add Create ScheduleDTO

+++ adapters/db/diagnostic.centre.sql.go
@@ -150,7 +150,7 @@ type Find_Nearest_Diagnostic_Centres_WhenRejectedRow struct {
 	AvailableTests       []string           `db:"available_tests" json:"available_tests"`
 	CreatedAt            pgtype.Timestamptz `db:"created_at" json:"created_at"`
 	UpdatedAt            pgtype.Timestamptz `db:"updated_at" json:"updated_at"`
-	DistanceKm           float64              `db:"distance_km" json:"distance_km"`
+	DistanceKm           int32              `db:"distance_km" json:"distance_km"`
 }
 
 func (q *Queries) Find_Nearest_Diagnostic_Centres_WhenRejected(ctx context.Context, arg Find_Nearest_Diagnostic_Centres_WhenRejectedParams) ([]*Find_Nearest_Diagnostic_Centres_WhenRejectedRow, error) {
@@ -325,7 +325,7 @@ type Get_Nearest_Diagnostic_CentresRow struct {
 	AvailableTests       []string           `db:"available_tests" json:"available_tests"`
 	CreatedAt            pgtype.Timestamptz `db:"created_at" json:"created_at"`
 	UpdatedAt            pgtype.Timestamptz `db:"updated_at" json:"updated_at"`
-	DistanceKm           float64             `db:"distance_km" json:"distance_km"`
+	DistanceKm           int32              `db:"distance_km" json:"distance_km"`
 }
 
 // Retrieves the nearest diagnostic centres based on latitude and longitude.
